Handle several collectors. Record per-tweet language.
We can now use several collectors, all storing in the same
database. The tweet table has an additonnal column called
'collector_id' referencing a new table 'collector' describing
collectors. The id of the collector must be specified in the
conf.json file.

The twitter-provided estimation of tweet language is now
saved in a new column 'lang_id' in table 'tweet'. It
references one of the languages listed in table 'languages',
that was already used to record the language specified in
user profiles. If twitter could not find the language,
we get 'und' instead of the language code.