Fix race condition in loop-collect.sh
The screen session could be started several times if collector.py
exited and loop-collect.sh was called during the 3 seconds delay.

We now look for the screen session instead of the collector.py
process.

We also remove the cron-restart.sh shell script and its entry in the
crontab, it is not useful anymore.